Move LFM Overlay from ACC to Common
