+++ Race Element.HUD.Common/Race Element.HUD.Common.csproj
@@ -29,6 +29,7 @@
         <PackageReference Include="EdCanHack.Unglide" Version="2.0.0" />
         <PackageReference Include="Riok.Mapperly" Version="4.1.1" />
         <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.1.0" />
+        <PackageReference Include="System.Speech" Version="9.0.0" />
         <PackageReference Include="TwitchLib.Client" Version="3.3.1" />
         <ProjectReference Include="..\Race Element.Data\Race Element.Data.csproj" />
         <ProjectReference Include="..\Race_Element.HUD\Race Element.HUD.csproj" />
